fix: correct float32 to float64 dtype in tuab_dataset for mypy compliance
- Changed output_data dtype from np.float32 to np.float64 in _load_edf_file
- Resolves mypy type error about incompatible return value types
- Maintains consistency with method return type annotation

+++ src/brain_go_brrr/infra/data/tuab_dataset.py
@@ -300,7 +300,7 @@ class TUABDataset(Dataset[tuple[torch.Tensor, int]]):
         data = raw.get_data()
 
         # Create output array with exactly 23 channels
-        output_data = np.zeros((len(self.STANDARD_CHANNELS), data.shape[1]), dtype=np.float32)
+        output_data = np.zeros((len(self.STANDARD_CHANNELS), data.shape[1]), dtype=np.float64)
 
         # Fill in available channels in the correct order
         for idx, ch_name in enumerate(self.STANDARD_CHANNELS):
